# src/swagger.schema.json
{"openapi":"3.0.0","paths":{"/json":{"get":{"operationId":"AppController_getSwaggerJson","parameters":[],"responses":{"200":{"description":""}}}},"/user":{"post":{"operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Users"],"security":[{"bearer":[]}]},"get":{"operationId":"UserController_find","parameters":[{"name":"limit","required":false,"in":"query","description":"How many in one page","schema":{"minimum":1,"maximum":25,"type":"number"}},{"name":"page","required":false,"in":"query","description":"Page number","schema":{"minimum":0,"type":"number"}},{"name":"order","required":false,"in":"query","schema":{"enum":["ASC","DESC"],"type":"string"}}],"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/user/me":{"get":{"operationId":"UserController_getMe","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserNotFoundException"}}}},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Users"],"security":[{"bearer":[]}]},"put":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/address/{id}":{"get":{"operationId":"AddressController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Address"],"security":[{"bearer":[]}]},"put":{"operationId":"AddressController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAddressDTO"}}}},"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Address"],"security":[{"bearer":[]}]},"delete":{"operationId":"AddressController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Address"],"security":[{"bearer":[]}]}},"/address":{"get":{"operationId":"AddressController_find","parameters":[{"name":"limit","required":false,"in":"query","description":"How many in one page","schema":{"minimum":1,"maximum":25,"type":"number"}},{"name":"page","required":false,"in":"query","description":"Page number","schema":{"minimum":0,"type":"number"}},{"name":"order","required":false,"in":"query","schema":{"enum":["ASC","DESC"],"type":"string"}}],"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Address"],"security":[{"bearer":[]}]},"post":{"operationId":"AddressController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAddressDTO"}}}},"responses":{"201":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Address"],"security":[{"bearer":[]}]}},"/notification-token/update":{"post":{"operationId":"NotificationTokenController_updateNotification","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateNotificationTokenDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationTokenEntity"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Notification Token"],"security":[{"bearer":[]}]}},"/notification-token/{id}":{"delete":{"operationId":"NotificationTokenController_deleteById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}},"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}}}},"tags":["Notification Token"],"security":[{"bearer":[]}]},"get":{"operationId":"NotificationTokenController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Notification Token"],"security":[{"bearer":[]}]},"put":{"operationId":"NotificationTokenController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateNotificationTokenDto"}}}},"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Notification Token"],"security":[{"bearer":[]}]}},"/notification-token":{"get":{"operationId":"NotificationTokenController_find","parameters":[{"name":"limit","required":false,"in":"query","description":"How many in one page","schema":{"minimum":1,"maximum":25,"type":"number"}},{"name":"page","required":false,"in":"query","description":"Page number","schema":{"minimum":0,"type":"number"}},{"name":"order","required":false,"in":"query","schema":{"enum":["ASC","DESC"],"type":"string"}}],"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Notification Token"],"security":[{"bearer":[]}]},"post":{"operationId":"NotificationTokenController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNotificationTokenDto"}}}},"responses":{"201":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationException"}}}},"401":{"description":""},"422":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TypeOrmExceptionFilter"}}}}},"tags":["Notification Token"],"security":[{"bearer":[]}]}}},"info":{"title":"Server","description":"","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:4000/api"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ValidationException":{"type":"object","properties":{}},"TypeOrmExceptionFilter":{"type":"object","properties":{}},"CreateAddressDTO":{"type":"object","properties":{"street":{"type":"string","example":"123 Dave St","description":"The city"},"houseNumber":{"type":"string","example":"123","description":"The city"},"city":{"type":"string","example":"Mississauga","description":"The city"},"district":{"type":"string","example":"Meadowvale Village","description":"The district"},"province":{"type":"string","example":"Ontario","description":"The province"},"country":{"type":"string","example":"Canada","description":"The country"},"postalCode":{"type":"string","example":"L5W 1K2","description":"The postal code"}},"required":["street","city","province","country"]},"CreateNotificationTokenDto":{"type":"object","properties":{"deviceId":{"type":"string","description":"unique device id"},"fcmToken":{"type":"string","description":"Fcm token"}},"required":["deviceId","fcmToken"]},"CreateUserDto":{"type":"object","properties":{"firebaseUID":{"type":"string","example":"Y0hYEkX9S3UDGoD4l961LxosZFj2"},"name":{"type":"string","example":"John Doe"},"email":{"type":"string","example":"john@gmail.com"},"dateOfBirth":{"format":"date-time","type":"string","example":"2023-11-08T04:07:46.969Z"},"gender":{"enum":["MALE","FEMALE","TRANSGENDER","GENDER_NEUTRAL","NON_BINARY","NONE"],"type":"string","description":"Male, Female etc. "},"address":{"type":"array","items":{"$ref":"#/components/schemas/CreateAddressDTO"}},"notificationTokens":{"type":"array","items":{"$ref":"#/components/schemas/CreateNotificationTokenDto"}}},"required":["firebaseUID","name","email","dateOfBirth","gender","address","notificationTokens"]},"AddressEntity":{"type":"object","properties":{"id":{"type":"string","example":"9a42511c-c4b3-4b51-98e7-0ca484760792"},"city":{"type":"string","description":"The city"},"district":{"type":"string","description":"The district"},"province":{"type":"string","description":"The province"},"country":{"type":"string","description":"The country"},"postalCode":{"type":"string","description":"The postal code"}},"required":["id","city","province","country"]},"NotificationTokenEntity":{"type":"object","properties":{"id":{"type":"string","example":"9a42511c-c4b3-4b51-98e7-0ca484760792"},"deviceId":{"type":"string"},"fcmToken":{"type":"string"}},"required":["id","deviceId","fcmToken"]},"UserEntity":{"type":"object","properties":{"id":{"type":"string","example":"9a42511c-c4b3-4b51-98e7-0ca484760792"},"firebaseUID":{"type":"string","example":"12345"},"name":{"type":"string","example":"Joe Doe"},"email":{"type":"string","example":"john@gmail.com"},"role":{"type":"number","example":"USER"},"dateOfBirth":{"format":"date-time","type":"string","example":"Oct 17 1997"},"gender":{"enum":["MALE","FEMALE","TRANSGENDER","GENDER_NEUTRAL","NON_BINARY","NONE"],"type":"string","description":"Male, Female etc. "},"address":{"example":"Mississauga","type":"array","items":{"$ref":"#/components/schemas/AddressEntity"}},"notificationTokens":{"example":{"deviceToken":"id","fcmToken":"id"},"type":"array","items":{"$ref":"#/components/schemas/NotificationTokenEntity"}}},"required":["id","firebaseUID","name","email","role","dateOfBirth","gender","address","notificationTokens"]},"UserNotFoundException":{"type":"object","properties":{}},"UpdateAddressDTO":{"type":"object","properties":{"street":{"type":"string","example":"123 Dave St","description":"The city"},"houseNumber":{"type":"string","example":"123","description":"The city"},"city":{"type":"string","example":"Mississauga","description":"The city"},"district":{"type":"string","example":"Meadowvale Village","description":"The district"},"province":{"type":"string","example":"Ontario","description":"The province"},"country":{"type":"string","example":"Canada","description":"The country"},"postalCode":{"type":"string","example":"L5W 1K2","description":"The postal code"}},"required":["street"]},"UpdateNotificationTokenDto":{"type":"object","properties":{"id":{"type":"string","description":"UUID"},"deviceId":{"type":"string","description":"unique device id"},"fcmToken":{"type":"string","description":"Fcm token"}}},"SuccessResponse":{"type":"object","properties":{}}}}}